add commo test to dotnetcore and dotnetframework to check for a bug
