+++ SauceExamples/Common/packages.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Appium.WebDriver" version="3.0.0.2" targetFramework="net461" />
+  <package id="Castle.Core" version="3.3.3" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
+  <package id="Selenium.Support" version="3.0.1" targetFramework="net461" />
+  <package id="Selenium.WebDriver" version="3.0.1" targetFramework="net461" />
+</packages>
