+++ SauceExamples/SauceExamples/AppiumMsTest.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Appium.WebDriver" Version="3.0.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Common.csproj" />
+  </ItemGroup>
+
 </Project>
